Update and rename omgimsoclose.js to periodic.beta.js
This is very close to what I had envisioned for the table framework. I finally got it so that all of the values in the draw function are increasing at a consistent rate. There is still one small bug with the mouse detection feature. When you hover over lithium, the element symbol increases in size, but the hydrogen symbol does as well. I'd like to make it so that only one element symbol increases in size at a time.

I have a separate block of JS code that demonstrates the mouse hover event action a little more effectively. However, that will need some work in order to be integrated properly into the table framework. You can see this block of code in the file "mouseHoverMore.js"

I'm going to take a break from the JS framework for a day or two, unless a useful idea comes to mind within that time. Meanwhile, I'm going to continue to practice HTML, particularly referencing CSS and JS files in an HTML doc. The javascript files for the periodic table are getting pretty large, for the sake of efficiency, I'm going to need to separate the table framework into chunks that can interact with each other via an HTML doc as they would in a single JS file. Also, I think that would make it easier to fix the issue with the cursor detection portion of the current table framework doc.

Additionally, I would like to start using VS code. I've been using editor.p5js.org in combination with sublime text until now. I've had some issues installing extensions that would allow me to write JS and Python code in sublime text. Some of the features Preston mentioned about VS code make it sound very appealing. The extension marketplace seems easier to use.